Citywatch rework
Comparing the stats between the roles of the Watchman/citywatch and the Manorguard/MAA, its clear that the watchmen were buffed for little reasoning.

The armour already is better than blacksteel, why then also make the physical stats and skills also better. These are untrained peasants meant to have a roof over their heads and food to eat. A slightly better position than the bog guards.

And yes, am awake that a MAA rework/respite is in the works, though IMO the only thing it will change is just buff the MAA to absolute hell, to the wish and joy of no one. No one asked for the watchmen to be buffed, and this change is meant to counteract the combination of armour that's better than blacksteel, with a stunmace. 

If nothing else, its a temporary change at best