Public Workshop (#1542)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Woodcutters were kicked from Artificers Guild HQ. Hunters are homeless.
Introducing... The Laborer's Workshop!


![image](https://github.com/user-attachments/assets/da87674b-b0bb-487c-90c8-79c039db8fc6)

Its located next to the Towner houses, in the outskirts of the city!


![image](https://github.com/user-attachments/assets/6d233d72-1edd-41fb-aa41-0b66d4e4c7d3)

It has everything needed for Woodcutters to work their wood!


![image](https://github.com/user-attachments/assets/09b57b4d-827b-4f99-b4d4-b8360eefb996)

And for hunters! To work their leather!


![image](https://github.com/user-attachments/assets/db7df441-f30a-45c4-8433-b1765cd6702c)

The small woodcutter house that was next to it was ruined. Preparing for
further deletion in the future!

The Workshop includes 2 carts, for the laborers to take their goods to
town to sell!

I also left some items to make it obvious to players how they should use
the rack and the grinder, if they don't yet know how to!

## Why It's Good For The Game

Gives Woodcutter and Hunter Towners a place to work!
Which is very nice and good. Towners need love.
Also further reinforces the idea of ruralization for Towners.