Allowed Rogue patron changes (#1548)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
The PR allows Rogues to have any patron instead of Xylix.
## About The Pull Request

The change allows Rogues to select any of the Divine Ten as their patron
God instead of only being able to select Xylix or inhumen Gods.

## Why It's Good For The Game

It feels unnecessarily restrictive to RP if the only god Rogues can
mechanically worship is Xylix. It makes sense that Xylix would be the
most common deity worshipped by Rogues, but I can see reasons why Rogues
would worship other deities while still taking up their trade. For
example, Swashbuckler is a variant of Rogue and a Swashbuckling
Zorro-styled vigilante would have Ravox as their patron since a
vigilante is about fighting injustice. It allows for more creative
freedom with backstories and I can't think of a downside to allowing
Rogues to be freeform with their patron.

---------

Co-authored-by: Gyrandola <pasta.frollagg@gmail.com>