Merchant key heat death (#1555)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Removed the 2 Shop keys and 1 Merchant key from the Merchant shop.

- Locked the door to that fucking thing I don't remember the name of
that takes items and turns them into gold

This goes along with #1552

"Buh buh buh Desdemona, if theres no merchant how will we make use of
this?!!??!! 😭 😭 😭 😭 😭 😭 😭 😭 "

YOU DON'T.
Do NOT. You do NOT. You DON'TN'T.

Shophand spwans with a shop key and can get a Merchant key from the
Goldface if theres no merchant.
Merchant spawns with both keys.

If there are no Merchant or Shophand you don't make use of their
services. Simple as.

## Why It's Good For The Game

The answer to "Why play the role if anyone can just break in and do it?"
Players should not be able to accomplish anything and everything by
themselves.