Physician's Apprentice
Another five second Carl addition, made at the request of a few folks.
We don't have sprites, so I'll let this sit for a bit, but it's fairly simple.
Adds an apprentice role for the court's Physician, permitting newer players to learn the medical system in a much more relaxed manner. Rather than being tossed into the deep end.