Revert "Prisoner Hermes (#1546)" (#1556)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This reverts commit 60870ba69a37aa05f19158c23f731c6f43620151.

Giving the prisoners a HERMES on their own cells quite litterally lets
them get mailed anything with a weight that's less than bulky, here's
pictures of an accopmlice with a full set of blacksteel armor and a whip
sending it to their prisoner friend, just in case you need proof you can
do it.
Accomplice POV:
![Captura de pantalla 2024-11-12
115746](https://github.com/user-attachments/assets/5c842b18-3c11-470c-bef0-03f7669a3bf9)
![Captura de pantalla 2024-11-12
115846](https://github.com/user-attachments/assets/2f41d65c-0935-4aeb-9d28-e691525ffc37)
![Captura de pantalla 2024-11-12
115925](https://github.com/user-attachments/assets/43431cfc-7c5a-476f-9152-07fd7fbf0bab)
Prisoner POV:
![Captura de pantalla 2024-11-12
120007](https://github.com/user-attachments/assets/f1509a99-4b8a-4f8f-a877-d143bc92b374)
![Captura de pantalla 2024-11-12
120057](https://github.com/user-attachments/assets/789df372-4966-4647-ae27-9362f2e7fe94)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Prisoners shouldn't be able to smuggle things into, not only the
dungeon, but their personal cells that easily. I get that they were
added to make them able to send mail and such and talk with other
people, and i might make a special hermes that doesn't let them recieve
items, but i hope this gets reverted before someone abuses it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->