Bandit migration (#1550)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds back bandits being able to respawn with waves in a reduced manner.
I also ended up talking with the person who had removed it instead of
reducing it, and he does not know why it was removed instead of being a
reduced migration.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Currently, bandits tend to die very quickly, it can be 40 minutes into
the round or an hour. But having some way that a few players (2 instead
of 4) can come back to make some action for the bog gate, or to
replenish their lines. Can do a lot of good. And this is even more with
the respawn changes, and the future bandit rework.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->