Grenzel Merc PQ Changes (#1563)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I raised the PQ requirements for both Grenzelhoft Mercenary and
Kommandant.
Mercenary: 2 to 15
Kommandant: 6 to 18

## Why It's Good For The Game

Grenzelhoft Mercenaries are an incredibly powerful role with no
allegiance. Their descriptions imply a loyalty to maintaining the
Guild's reputation, but this is usually not followed. I am increasing
the PQ requirement to decrease the amount of fraggers and LRP players
choosing this role.