Runelock Pistol Changes (#1564)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I wanted to make runelocks smoother.
Ended up refractoring gun code a bit.

- Runelocks are now fire instantaneously, skill dependent on hunting for
reloading speed and spread reduction (not that it matters much because I
didn't increase the spread enough for it to be noticeable)

- They also have a more fitting sound now and produce smoke.

- Removed De-cocking because it served no function and caused the player
to have to reload the gun for 10 seconds again if they accidently
pressed use.

- Changed a some descriptions and item names to do away with the "magic"
theme.

- Changed Projectile sprite to make it hopefully more noticeable

- Removed Runelock as Metalizer result from Crossbow as per Useroths
request.

- Created some vars and moved vars around to make the code more reusable
with little effort in the future.

Little to no changes noticeable in game aside from instant shooting and
smoke:
***For some reason I had forgotten to change the sound here***


https://github.com/user-attachments/assets/9cd3141c-2042-4771-938d-2b48a8d5f4c0


## Why It's Good For The Game

I think the gun feels better to wield now.
Its also more in line with server design direction.