Fixes butchering skeleton limbs and zombie limbs being butcherable (#1578)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Disables the ability to butcher zombie and skeleton limbs, puts a
warning in chat when attempting to
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Butchering zombies and skeletons was irrational, could make butchering
zombies give rotted meat for pestrans but...
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->