Removes Bathoa's Curse From Lovepotion (#1579)
## About The Pull Request

This removes the Bathoa's curse hook from lovepotions.

## Why It's Good For The Game

I have seen absolutely nothing but player complaints from the addition
of bathoa's curse to the lovepotion item in PR 1486
[](https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1486). Lovepotion
is the yandere roulette potion, not the cumbrained blind crippled slut
potion. Adding bathoa's curse to it- when the potion bypasses Defiant,
and with the current the trainwreck blinding and crippling
implementation, was an overreach and has gotten nothing but negative
player feedback. If you want a cumbrained slut potion, get me a sprite
and I'll make a pr for you and add it to the merchant. However
lovepotion is for rolling the _yandere obsession_ roulette You _might_
get an obedient obsessed bottom. You might get an aggressive kidnapping
inclined yandere, you might get one that wants to cook and eat you so
you'll never be apart ever again. It depends wholly on the victim
character's interpretation. Adding the cumbrain affliction takes away
from potential interesting rp and scenes, on top of crippling someone,
even a defiant character that did not consent to erp mechanics with you,
for an entire ass round. Thus, removing that.

## Testing



https://github.com/user-attachments/assets/eab47181-a1c0-4bea-ae83-15f3e0ca4c8a