Re-arranges order of Heartfelt wave (#1588)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Re-arranges order of party members lost for the heartfelt wave.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The entire reason for my last PR allowing both men and woman as lord and
lady of heartfelt, is that getting someone to PICK the lady of heartfelt
is like pulling fucking teeth. Even if we have 5 people rolling for the
heartfelt wave, No picks the lady, and then we're left with losing over
half the roles from the wave because of it.

Hopefully this adjustment will make heartfelt waves more likely to
succeed in the lobby.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->