Cutting Stump is no longer Ass (#1576)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Now chopping a stump works the same way as chopping a lot:


![image](https://github.com/user-attachments/assets/4684dfc3-b4ff-4cc8-81ff-e591861d3491)


## Why It's Good For The Game

As of now cutting a stump only delays the player for no real gain, which
lead to a lot of stumps everywhere.
Players would rather have to push the cart with them and climb on top of
the stump to store the logs in their cart and do this for every tree
than to actually remove the stump.
Since players liked this woodcutting mechanic, I decided to expand it to
stumps to solve the problems highlighted above.