Makes tent walls unclimbable (#1593)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Does as it says, makes tent walls unclimbable. So no more building 4
floor high tent walls to go above the keep walls for bandits. This also
prevents them from being climbed if a wall ladder is built too. So don't
worry about them building one of them either.

![Captura de pantalla 2024-11-14
170510](https://github.com/user-attachments/assets/c19ea752-26f7-4e0e-afe7-09a25200a46b)
![Captura de pantalla 2024-11-14
170536](https://github.com/user-attachments/assets/a06f2197-6dc8-4bd6-a26e-53741b5cd6fd)

Also did it mostly since Gyran and MickDick1 asked for it in
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1584 and
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1583
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Same as said below, if you want to build a staircase high enough to go
through walls that are specifically made to not be moved above of, go
get a proper mason.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->