Fix for Smelter Bug (#1594)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

My previous PR caused a bug that basicly made steel impossible to make.

While that was extremely amusing to me and caused a very wicked event
where people were scrapping for steel everywhere, its not what I
intended.

This fixes it.

Made 6 steel this time:

![image](https://github.com/user-attachments/assets/a7ca38c8-cc34-4c3c-b108-5d91f400689d)

## Why It's Good For The Game

I fucked.