Mimic & Poltergeist
Adds both of the creatures mentioned in the title.

Functionality:

Mimic pretends it's a normal chest, up until the point someone comes close to it. When activated, it'll unfold and reveal itself. Upon death, it'll drop whatever was placed into it via mapping, or whatever is on its tile when spawned. There exists a single mimic(by chance) on the map, and it's in the wizard's tower, replacing the coffin that had spice in it, as a 50/50 between a mimic or a chest spawn. Waiting on map stuff to be done before actually mapping these in proper.

Poltergeist is something I pulled near 1:1 off of what I did years ago for Stalker 13. It's supposed to be a floating skull, like a D&D demilich, but for now we're using a temporary sprite. The unfinished sprite is in the code, but, of course, is unused and marked as 'new' for now.
For its functionality, it floats around, makes weird noises and throws whatever it can see at players. Like a violent haunting, generally causing a mess. It's near unhittable, though with incredibly low health. Use Churn Undead to get rid of it, or swing when it becomes temporarily visible for a chance to do something.
There also exists an event for a single time per round, during an omen as a 10% chance, that it'll spawn a handful of them around the map.

I thought these were neat enough to add. That's about it.
Mimics serve as a way to punish people rushing loot spawns, for better or worse.
Poltergeists are unique fluff mobs that can spook people.