Re-Enables Seelie Jester (#1602)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Lets Seelie be jesters again.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Looks like this was broken, probably by the efforts to stop seelie from
being bandits. Here's a screenshot to show it working after the change..


![image](https://github.com/user-attachments/assets/7637c1b2-7c2b-49a2-a40b-715d0291b55d)


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->