Global spawn protection (#1533)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Makes every single role get spawn protection upon joining. 

- [x] tested

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

With the new changes, every latejoin role spawns outside the Bastion, it
is thus necessary for everyone to get the same onspawn protection
previously reserved for adventurers and pilgrims.