removes the Nypmhomatic "special" (#1607)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->It is not fun or unique and it does nothing but
mechanically hinder you.

## Why It's Good For The Game
This is what you see 24/7

![image](https://github.com/user-attachments/assets/1204ec97-0bcb-406a-87c2-8f6eed6d95a1)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Gyrandola <pasta.frollagg@gmail.com>