Stun Mace Return (#1510)
Makes stunmaces drain stamina and fatigue. When a target is at no
stamina, it acts like it used to, electrocuting and stunning the target.
The mace does 6 damage when active, 25 when off.
Also gives the guards gorgets. Because they used to have neck
protection.
## Why It's Good For The Game
The mechanic was removed temporarily pending rework. I'd argue this is
technically a rework.