Gives malum templars warhammers. (#1613)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Gives Malum templars warhammer and maces 4, every other templar gets
swords 4 and bastard sword
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

https://discord.com/channels/1240735983276654733/1258555111005880520/1307806572994625577
"Why it's good for the game: Malum is all about hammers, especially in
the clerical magic spells. It's more fitting, as he represents smiths
and other craftsfolk. The weapon focus provides a flavor to the
different faith the templar has."

https://discord.com/channels/1240735983276654733/1258555111005880520/1307856823239315456
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->