Rogue nerf Part 2 (#1618)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR: https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1615 tried to
give a (deserved) nerf to rogues, but they went a little bit overboard
with the nerf, giving swashbucklers a total of 0 overall stat points for
swashbucklers and 2 for thieves. For comparison, the bard, the least
combat-centric class of all adventurers gets 4 overall.


![image](https://github.com/user-attachments/assets/343db9d6-5a39-4592-a23c-ad138732d9d9)

![image](https://github.com/user-attachments/assets/240757eb-0f66-4cb7-99b2-17d42d093737)

It also removed their skill to wrestle which would make them defenseless
against anyone that was able to get a grab on them, since they have
already low STR. (All combat classes have at least apprentice wrestling
to avoid getting beat by a single grab. Even the mage.)

This PR also lowers the thief's sword skill since they're supposed to be
the "non-combat" version of rogues, besides, they already get great
stats on everything else.

On another note, this PR also nerfs the dodge expert trait from giving
you 15% chance per point to a 14% chance, it's a very low nerf, but it
should make hitting rogues with high very SPD when you've got high skill
easier. So hopefully it'll be harder to get a 90% dodge chance against a
knight.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

This just gives them again a reasonable amount of stat points and the
wrestling skill again, while keeping all the other nerfs applied on the
last PR the same.

Rogues deserved a nerf. Yes. But that nerf from the last PR was *way*
too much. Hopefully this is sliiiiiightly more balanced so it's not as
overpowered as they were before the nerf and not as underpowered as they
were after it.

Feel free to do any suggestions for more balance.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->