Bandits at it again - Hordemaster moved in (#1606)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Ports bandit hordemaster from [ BANDIT DLC PT 2 - PRAYING FOR NO MAP
CONFLICTS EDITION #659
](https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/659)

![image](https://github.com/user-attachments/assets/e609beeb-432c-43bf-880b-c590083f7d1c)


![image](https://github.com/user-attachments/assets/d6b3bd09-fb74-445a-8437-43f1c21151ca)

And a few sacks were added in

![image](https://github.com/user-attachments/assets/d4f99664-10a6-40ba-9cac-05dee7c21e39)


The balance was slightly changed, so the bandits have to work for their
triumph instead of the original PR starting with 150 > was changed to 10

The PR will most likely be completely whack in terms of balance, but as
this was mostly a straight port. Then I choose not to touch much of it
so that I can hear feedback in a TM.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->


## Why It's Good For The Game
Remove the awful items being dropped that are boring and annoying to
work with and let the bandits choose from a list of items for what they
want to buy, from a general list and class level.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->