Makes objects more resistant to fire (#1590)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
partially reverts the changes to fire and changes the mod from 5x damage
to 2x damage
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1126
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1104

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Being naked but otherwise fine because your clothes burnt off is bad, I
would rather fire only deal a fraction of the health of clothing and it
needing repairing rather than having being naked but alive be a high
possibility.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->