Goblin Rework(Ish) (#1166)
## About The Pull Request
- - -
Player Facing:
- Goblin roles re-enabled once again, this time with a bit of a
facelift. You're still required to have Graggar as a patron.
 - Goblin Guard and Chief are still gender locked.
 - Goblin Chef and Smith are now free to play for both genders.
- Goblin Rabble added. This is the 'commoner' equivalent for the Goblin
encampment. They're not given access to the interior by default, having
to wait for it to be opened. This, alongside Smith and Chef, are two
introductory roles.
- Player Goblins should now actually be able to wear their armour.
Rejoice?
- The Goblins have travel tiles, allowing them to go from their fort to
the sewers, as was requested.
- - -
Other:
- Goblin camp deleted from the game files. Didn't make sense and had
only served to uglify the file structure even worse than it had already
been.
 - Goblin Camp map pass done by @NikolayByondPlayer.
- - -

## Why It's Good For The Game
Objectively, I couldn't tell you one way or another. This was done on
request, for which I figured I'd give it a bit of a lookover. It's still
really rough, especially with how it had been added in the first place,
but it should be better than prior. The issues of fragging can be solved
with administrative action.

---------

Co-authored-by: Gyran <62253058+Gyrandola@users.noreply.github.com>