Updates the Description of Love Potions and the Cautions Associated with Love Potions (#1624)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Gives love potions a description (it had none) and re-flavours the
cautions.

## Why It's Good For The Game

Currently, the feelings are not defined. While that is good for roleplay
since not everyone is capable of feeling love, it should be a little
more guided. What you did get text-wise is short too. I recall an
incident where a bandit silently fed the Duke (King at that time) a love
potion in court during a chaotic moment, but he missed the text you get
and he ended up executing her. With these changes it will be longer, so
they won't be missing it. It also removes mention of developing feelings
for people the potion-giver likes. Why should someone like the Duchess
and her children if the Duke feeds them a love potion, you are now
competing with them for the Duke's affections. While I understand love
potions apply the faction tags of the potion-giver to the recipient, it
only has mechanical effects for simple mobs (that I am aware of). If
people want, they can still develop feelings for others the potion-giver
likes. Otherwise, they shouldn't be forced to. Let there be risks
associated with love potioning bandits and killers.

For the caution of trying to drink the potion yourself I also gave a
hint to try and feed it to someone else.

Photo evidence:

![When
Fed](https://github.com/user-attachments/assets/3505e360-d9f7-40ea-8af3-03f6e4115acb)

![Love Potion Descruption and
Cautions](https://github.com/user-attachments/assets/a8cf3a51-6408-4cff-9c60-c4b94495dea0)