Zizoid Beast Invulnerability | End Message Touchup
We're not meant to be able to kill the Zozoid ascension creature, but for some reason, it's a big ol' joke. This makes it immune to damage, as had been intended with its health set so damn high. Ascension is meant to be a big deal.
This PR is probably overkill, but I'm lazy. We'll probably not merge this, but it's here just in case.

As an aside, it also touches up the end round message to be more clear. IMO, at least.