Druid Buff | Scomm Ring Icon
Provides druids a hefty buff.

Adds them being friendly with the following:
 - Spiders
 - Rats
 - Volves
 - Moles

They're also provided a walking stick by default, with randomized skill in polearms between 1 or 2.

Additionally, to allow them to interact easier with WWs (and others, RP wise), they've also been given the beast language, as they're servants of Dendor and should probably be able to communicate with his creations.

Finally, it makes them immune to being ambushed by walking over tiles.

Not related to this PR's main focus, but snuck in, is scomm rings getting a temporary sprite by way of using the scommstone icon. It doesn't have its two special states set, however, and is one icon only. It's not proper at all, but it's something to look at that makes SOME sense, as opposed to it being invisible.