Allows druid to use beasttongue (#1640)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The Verevolf zodomites wouldn't shut up about it.

I don't care about this PR at all and wont' be responding to balancing.

If they want it merged, they need to argue for it.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Easier to fuck verevolfs for druids ig
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->