Zizoid Beast Invulnerability | End Message Touchup (#1638)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
We're not meant to be able to kill the Zozoid ascension creature, but
for some reason, it's a big ol' joke. This makes it immune to damage, as
had been intended with its health set so damn high. Unless I'm mistaken?
I'm pretty tired as I write this. Ascension is meant to be a big deal,
in any case. This PR is probably overkill, but whatever.

As an aside, it also touches up the end round message to be more clear.
IMO, at least.

## Why It's Good For The Game
Cult ascension was more of an inconvenience, with the creature killed
easily. The mood debuff was a bigger issue, by comparison. This should
rectify that by throwing it well into the "OH GOD OH FUCK" realm of
being broken, as was the intent originally, unless I'm mistaken.

The message was also not clear. This fluffs it up a bit and straight up
states that it's the end in fifteen minutes, not just 'fifteen minutes
remain'.