Druid Buff | Scomm Ring Icon (#1639)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Provides druids a hefty buff.

Adds them being friendly with the following:
 - Spiders
 - Rats
 - Volves
 - Moles

They're also provided a walking stick by default, with randomized skill
in polearms between 1 or 2.

Not related to this PR's main focus, but snuck in, is scomm rings
getting a temporary sprite by way of using the scommstone icon. It
doesn't have its two special states set, however, and is one icon only.
It's not proper at all, but it's something to look at that makes SOME
sense, as opposed to it being invisible.

## Why It's Good For The Game
Druids are kind of a forgotten role. I 'unno. This was at the request of
a few people, with suggestions from players.