Hordemaster Foresworn Bevor Fix (#1650)
## About The Pull Request

Fixes Hordemaster Bevor buying for Foresworn and adds the Hound Mask as
an additional purchasable option.

## Why It's Good For The Game

Fixes the issue where Foresworn would get scammed for hundreds of favor
because a single typo was made in the code(Classic).

As an example of what the issue was for those unaware, Foresworn were
able to purchase a Bevor for on avg a 100 favor. As a result of the typo
in the spawn code, you would spend the favor, it would be removed from
your balance and you would get nothing. (Then the Hordemaster would go
"Thank you~", ruining your day because that greedy fat fucker just stole
your shit.)

Also added the Steel Hound Mask as a purchasable option for everyone
because style is important. Same cost as the regular Steel Mask.