Some hordemaster cleanup (#1652)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Hopefully, the last time that we will see the need to touch this piece
of code in terms of items.

![image](https://github.com/user-attachments/assets/4c594a89-b2da-4705-adb6-5e281510a1a8)

![image](https://github.com/user-attachments/assets/8c31d985-f805-4b52-acce-7624345d452d)

![image](https://github.com/user-attachments/assets/822fc887-02b0-479e-b3ac-6c32f0ec7209)

Added: stuff - Since the things section was getting too much stuff in
it... Hah
Added: a new more stealthy armored hood, for people who want to play
Robin Hood
Added: a lot of clothes variety, masks
Added: missing shields for the foresworn
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More clothes variety gives more access for bandits to get in stealthily.
And the masks are there for more variety, and more !fun!
You can now also pick items without having to scroll down all the time.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->