Excidium & Guild clerk rework (#1656)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Dear rockhill.

Thank you for choosing to pay for the upgrades and say hello to yout
*new* excidium machines! Here's a list of features that this upgrade
comes with that should hopefully make it's usage more common and easier
for you:

- Custom bounties!
Do you know how other places tend to have a little board with tasks
people ask help with? Anyways, now you can use the excidium for that!
(The creators probably intended it for darker purposes, but you're free
to use it as you may.)
Now you can create custom bounties that everyone will be able to see and
do. All you need to do is talk to the guild clerk after you're done with
it so they can verify you actually did it and then you'll be good to go.
You should probably get some proof that you *actually* completed the
task before going to bother him, don't want to waste his time.

- Non-lethal bounties!
Now when choosing a bounty you can decide to have *some* mercy on the
poor fella you're ruining the day of, you can decide to give the guild
clerk the option to mark the bounty as completed *without* having to
deliver the head of the person. You should probably use this if you want
to let them have a chance to be captured alive or anything else other
than beheaded.
It should be clarified though, anyone can still get the reward on their
head by beheading them, this just simply gives them a *chance* to not
get beheaded. You should probably offer a greater reward if they're
brought alive if you'd like it that way, or make a custom bounty if
you'd like to be sure that they're 100% not killed.

Both of these bounties will have a number attached to them that'll be
used when trying to mark the bounty as completed.

Example of the three types of bounties that can be placed: (Ignore the
misplaced bounty number on the first one, that was already fixed)

![image](https://github.com/user-attachments/assets/9332fc06-f93e-45cc-8ef9-918f928848fb)

![image](https://github.com/user-attachments/assets/d13ded76-45e9-4131-a6f5-6b4bd9a46b20)

- Guild clerk rework
Now the guild clerk is it's own role, with their own slots, separated
from the mercenaries. So now you can see if there's already a guild
clerk around and don't have to join as a mercenary just to check.
Other than that they now have the new responsibility of managing the
excidium! They'll have to check if bounties have been properly completed
and hand out the rewards for when they are, as they're the only ones
allowed and trusted by the guild to properly hand out the rewards.
You may also have the task of informing th eemployer depending on which
one it is, if someone puts up a task of "Bring me 5 apples" and someone
brings you, the clerk, 5 apples, you should give them the reward, and
then try to notify (HERMES is a good thing for this) the employer that
the apples have been delivered. Or at least ask for the person
withdrawing the reward to bring a little note signed that says "Yeah i
got the apples" to make sure they're not lying.

Silly merchant. You don't have access!

![image](https://github.com/user-attachments/assets/e90e7a48-f9f8-472a-90ed-a6fa4ef5ed15)

Guild clerk now has something to do!

![image](https://github.com/user-attachments/assets/d823185d-d8ca-434e-b095-466aff5e55c6)

![image](https://github.com/user-attachments/assets/cd41d9f5-adf5-4350-96b2-b9ca3386a966)

![image](https://github.com/user-attachments/assets/d6470328-86b6-4eb5-bd73-d75be58618ef)

![image](https://github.com/user-attachments/assets/fcb40898-3729-4bee-abe7-38461ec50f97)

![image](https://github.com/user-attachments/assets/fc443032-743f-417c-aeab-09d6067c2d66)

- Excidium taxes now raise and get added to the treasury depending on
the current kindom tax!

Before we had a flat and default 10% that went to the treasury, but
after some complaints from the duke (That may or may not have been
encourage by the steward) we've decided to make the tax equal to the one
that is set on the kingdom. So if the duke decides to set the tax rate
at 50%, 50% of the bounty money will go to the king. Hopefully this'll
be a good enough benefit for them to leave the machine functioning on
their town.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More things for the guild clerk to do, more uses for the excidium, more
types of bounties, non-lethal bounties, more ways to interact with other
people, more ways to pay for tasks and let them be known that isn't just
yelling through the SCOM!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->