Bandit roll - Now able to be chosen in time. (#1653)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The bandit roll was a bit hard to catch for a lot of people since it did
not have the waves as other rolls do
This will give a window for bandits to get the roll from 1 min to 2 mins
for two people, and 2 mins and 30 seconds for one person
With the new musketeer roll, I added a little more weight, which was in
line with the original weight it had before.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bandits give drama and combat to people. And when all the bandits are
dead, new blood should come in to make it more fun, so the round does
not become incredibly stale at the one-hour mark.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->